cpu/atmega_common: Extract clock definitions
Refactor clock definitions from cpu.h to clock.h.
